// src/models/Exigencia.tsx
// Generated by https://quicktype.io

export interface Exigencia {
    id:                   number;
    metaDias:             string;
    observacao:           string;
    oportunidadeId:       number;
    status:               string;
    tipoDataMeta:         number;
    tipoExigenciaId:      number;
    tipoUsuarioClienteId: number;
    titulo:               string;
}