// src/models/TipoUsuarioResponse.tsx
// Generated by https://quicktype.io

export interface TipoUsuarioResponse {
    id:        number;
    descricao: string;
    status:    string;
    charter:   number;
}

// export enum Descricao {
//     Estratégico = "ESTRATÉGICO",
//     Financeiro = "FINANCEIRO",
//     Operação = "OPERAÇÃO",
//     Rede = "REDE",
// }

// export enum Status {
//     A = "A",
// }